Rename `mixins` to `hf_hub`
